+++ weblate/settings_docker.py
@@ -1269,6 +1269,7 @@ WEBLATE_ADDONS = [
     "weblate.addons.removal.RemoveSuggestions",
     "weblate.addons.resx.ResxUpdateAddon",
     "weblate.addons.cdn.CDNJSAddon",
+    "weblate.addons.cdn.CDNFilesAddon",
     "weblate.addons.webhooks.WebhookAddon",
     "weblate.addons.webhooks.SlackWebhookAddon",
     "weblate.addons.fedora_messaging.FedoraMessagingAddon",